Introduce ZlsThumbnailProvider
ZlsThumbnailProvider is a dll that generates thumbnail images for zls files.
It operates within File Explorer and extracts a thumbnail image from the
draping board in a zls file. These thumbnail images serve as previews,
allowing artists to quickly identify the contents of a zls file without the
need to open it in an editor like z-weave.

For more technical details, plase refer the official Microsoft documenation
available at the following like:
https://learn.microsoft.com/en-us/windows/win32/shell/thumbnail-providers