Delete or replace dll after reboot
On some versions of Windows, ZlsThumbailProvider.dll can not be deleted or
replaced before restarting the system if it has been registered.

This commit teaches the installer to defer deletion or replacement of the
dll after system reboot, if necessary.

According to the manual, administrative privileges are required to use the
uninsrestartdelete or restartreplace flags.

See https://jrsoftware.org/ishelp/index.php?topic=filessection